When deleting a task with locations, removes them from Geofence too.

+++ app/src/main/java/carlos/alves/todotaskreminder/deleteTasks/DeleteTasksListViewModel.kt
@@ -1,8 +1,10 @@
 package carlos.alves.todotaskreminder.deleteTasks
 
+import android.content.Context
 import androidx.lifecycle.ViewModel
 import carlos.alves.todotaskreminder.ToDoTaskReminderApp
 import carlos.alves.todotaskreminder.adapters.TaskObject
+import carlos.alves.todotaskreminder.notifications.LocationReminderService
 import carlos.alves.todotaskreminder.settings.SettingsConstants.*
 
 class DeleteTasksListViewModel : ViewModel()  {
@@ -12,6 +14,8 @@ class DeleteTasksListViewModel : ViewModel()  {
     private val onLocationRepository = ToDoTaskReminderApp.instance.onLocationRepository
     private val onlineTaskRepository = ToDoTaskReminderApp.instance.onlineTaskRepository
     private val settingsRepository = ToDoTaskReminderApp.instance.settingsRepository
+    private val locationRepository = ToDoTaskReminderApp.instance.locationRepository
+    private val locationReminderService = LocationReminderService.instance
 
     private lateinit var tasksNames: ArrayList<String>
 
@@ -24,11 +28,24 @@ class DeleteTasksListViewModel : ViewModel()  {
 
     fun generateTaskObjectList() = ArrayList(tasksNames.map { TaskObject(it) })
 
-    fun deleteTask(taskName: String) {
+    fun deleteTask(context: Context, taskName: String) {
         val taskId = taskRepository.getTaskId(taskName)
-        onLocationRepository.deleteOnLocationsByTaskId(taskId)
+        removeFromOnLocationsRepository(context, taskId)
         dateTimeRepository.deleteDateTime(taskId)
         onlineTaskRepository.deleteOnlineTaskByTaskId(taskId)
         taskRepository.deleteTask(taskId)
+    }
+
+    private fun removeFromOnLocationsRepository(context: Context, taskId: Int) {
+        val onLocations = onLocationRepository.getOnLocations(taskId)
+        if (onLocations.isEmpty()) { return }
+        onLocationRepository.deleteOnLocationsByTaskId(taskId)
+        onLocations.forEach { removeFromGeoFence(context, taskId, it.locationId) }
+    }
+
+    private fun removeFromGeoFence(context: Context, taskId: Int, locationId: Int) {
+        val location = locationRepository.getLocationById(locationId)
+        val geofenceId = "$taskId:${location.name}"
+        locationReminderService.removeLocationFromGeofence(context, geofenceId)
     }
 }
